1.4 Add LargePayload to VDM.
fix #3132

+++ include/library/spdm_requester_lib.h
@@ -1053,12 +1053,12 @@ libspdm_return_t libspdm_vendor_send_request_receive_response(
     uint16_t req_standard_id,
     uint8_t req_vendor_id_len,
     const void *req_vendor_id,
-    uint16_t req_size,
+    uint32_t req_size,
     const void *req_data,
     uint16_t *resp_standard_id,
     uint8_t *resp_vendor_id_len,
     void *resp_vendor_id,
-    uint16_t *resp_size,
+    uint32_t *resp_size,
     void *resp_data);
 
 #endif /* LIBSPDM_ENABLE_VENDOR_DEFINED_MESSAGES */
