+++ unit_test/fuzzing/test_responder/test_spdm_responder_vendor_cmds/vendor_cmds.c
@@ -35,9 +35,9 @@ libspdm_return_t libspdm_vendor_response_func_test(
     uint16_t req_standard_id,
     uint8_t req_vendor_id_len,
     const void *req_vendor_id,
-    uint16_t req_size,
+    uint32_t req_size,
     const void *req_data,
-    uint16_t *resp_size,
+    uint32_t *resp_size,
     void *resp_data)
 {
     return LIBSPDM_STATUS_SUCCESS;
